+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -11,9 +11,9 @@ public interface UserStorage {
 
     int update(User user);
 
-    Optional<User> get(int userId);
+    Optional<User> get(Long userId);
 
     List<User> getAll();
 
-    boolean delete(int userId);
+    boolean delete(Long userId);
 }
